implement cubit for state management and restructure onboarding page components
